Starting the armi.testing modules
