+++ armi/physics/neutronics/tests/test_crossSectionTable.py
@@ -21,8 +21,8 @@ from armi.physics.neutronics.isotopicDepletion import (
 )
 from armi.physics.neutronics.latticePhysics import ORDER
 from armi.reactor.tests.test_blocks import loadTestBlock
-from armi.reactor.tests.test_reactors import loadTestReactor
 from armi.settings import Settings
+from armi.testing import loadTestReactor
 from armi.tests import ISOAA_PATH
