+++ armi/reactor/converters/parameterSweeps/tests/test_paramSweepConverters.py
@@ -16,14 +16,14 @@
 import os
 import unittest
 
-from armi.tests import TEST_ROOT
+from armi.physics.neutronics.settings import CONF_EPS_FSPOINT
 from armi.reactor.converters.parameterSweeps.generalParameterSweepConverters import (
     NeutronicConvergenceModifier,
     ParameterSweepConverter,
     SettingsModifier,
 )
-from armi.reactor.tests.test_reactors import loadTestReactor
-from armi.physics.neutronics.settings import CONF_EPS_FSPOINT
+from armi.testing import loadTestReactor
+from armi.tests import TEST_ROOT
 
 
 THIS_DIR = os.path.dirname(__file__)
