+++ armi/reactor/converters/tests/test_meshConverters.py
@@ -17,7 +17,7 @@ import math
 import unittest
 
 from armi.reactor.converters import meshConverters, geometryConverters
-from armi.reactor.tests.test_reactors import loadTestReactor
+from armi.testing import loadTestReactor
 from armi.tests import TEST_ROOT
