+++ armi/reactor/tests/test_blocks.py
@@ -37,13 +37,12 @@ from armi.reactor import blocks, blueprints, components, geometry, grids
 from armi.reactor.components import basicShapes, complexShapes
 from armi.reactor.flags import Flags
 from armi.reactor.tests.test_assemblies import makeTestAssembly
-from armi.reactor.tests.test_reactors import loadTestReactor
-from armi.reactor.tests.test_reactors import TEST_ROOT
-from armi.tests import ISOAA_PATH
+from armi.testing import loadTestReactor
+from armi.tests import ISOAA_PATH, TEST_ROOT
 from armi.utils import hexagon, units, densityTools
 from armi.utils.directoryChangers import TemporaryDirectoryChanger
-from armi.utils.units import MOLES_PER_CC_TO_ATOMS_PER_BARN_CM
 from armi.utils.units import ASCII_LETTER_A, ASCII_LETTER_Z, ASCII_LETTER_a
+from armi.utils.units import MOLES_PER_CC_TO_ATOMS_PER_BARN_CM
 
 
 NUM_PINS_IN_TEST_BLOCK = 217
