+++ armi/reactor/tests/test_reactors.py
@@ -41,11 +41,11 @@ from armi.reactor.flags import Flags
 from armi.reactor.spentFuelPool import SpentFuelPool
 from armi.settings.fwSettings.globalSettings import CONF_ASSEM_FLAGS_SKIP_AXIAL_EXP
 from armi.settings.fwSettings.globalSettings import CONF_SORT_REACTOR
-from armi.tests import ARMI_RUN_PATH, mockRunLogs, TEST_ROOT
+from armi.testing import loadTestReactor, reduceTestReactorRings  # noqa: F401
+from armi.tests import mockRunLogs, TEST_ROOT
 from armi.utils import directoryChangers
 
-THIS_DIR = os.path.dirname(__file__)
-TEST_REACTOR = None  # pickled string of test reactor (for fast caching)
+_THIS_DIR = os.path.dirname(__file__)
 
 
 def buildOperatorOfEmptyHexBlocks(customSettings=None):
@@ -128,95 +128,6 @@ def buildOperatorOfEmptyCartesianBlocks(customSettings=None):
     o.r.core.add(a)
     o.r.sort()
     return o
-
-
-"""
-NOTE: If this reactor had 3 rings instead of 9, most unit tests that use it
-go 4 times faster (based on testing). The problem is it would breat a LOT
-of downstream tests that import this method. Probably still worth it though.
-"""
-
-
-def loadTestReactor(
-    inputFilePath=TEST_ROOT,
-    customSettings=None,
-    inputFileName="armiRun.yaml",
-):
-    """
-    Loads a test reactor. Can be used in other test modules.
-
-    Parameters
-    ----------
-    inputFilePath : str, default=TEST_ROOT
-        Path to the directory of the input file.
-
-    customSettings : dict with str keys and values of any type, default=None
-        For each key in customSettings, the cs which is loaded from the
-        armiRun.yaml will be overwritten to the value given in customSettings
-        for that key.
-
-    inputFileName : str, default="armiRun.yaml"
-        Name of the input file to run.
-
-    Returns
-    -------
-    o : Operator
-    r : Reactor
-    """
-    global TEST_REACTOR
-    fName = os.path.join(inputFilePath, inputFileName)
-    customSettings = customSettings or {}
-    isPickeledReactor = fName == ARMI_RUN_PATH and customSettings == {}
-
-    if isPickeledReactor and TEST_REACTOR:
-        # return test reactor only if no custom settings are needed.
-        o, r, assemNum = cPickle.loads(TEST_REACTOR)
-        o.reattach(r, o.cs)
-        return o, r
-
-    cs = settings.Settings(fName=fName)
-
-    # Overwrite settings if desired
-    if customSettings:
-        cs = cs.modified(newSettings=customSettings)
-
-    if "verbosity" not in customSettings:
-        runLog.setVerbosity("error")
-
-    newSettings = {}
-    cs = cs.modified(newSettings=newSettings)
-
-    o = operators.factory(cs)
-    r = reactors.loadFromCs(cs)
-
-    o.initializeInterfaces(r)
-
-    o.r.core.regenAssemblyLists()
-
-    if isPickeledReactor:
-        # cache it for fast load for other future tests
-        # protocol=2 allows for classes with __slots__ but not __getstate__ to be pickled
-        TEST_REACTOR = cPickle.dumps((o, o.r, o.r.p.maxAssemNum), protocol=2)
-
-    return o, o.r
-
-
-def reduceTestReactorRings(r, cs, maxNumRings):
-    """Helper method for the test reactor above.
-
-    The goal is to reduce the size of the reactor for tests that don't neeed
-    such a large reactor, and would run much faster with a smaller one.
-    """
-    maxRings = r.core.getNumRings()
-    if maxNumRings > maxRings:
-        runLog.info("The test reactor has a maximum of {} rings.".format(maxRings))
-        return
-    elif maxNumRings <= 1:
-        raise ValueError("The test reactor must have multiple rings.")
-
-    # reducing the size of the test reactor, by removing the outer rings
-    for ring in range(maxRings, maxNumRings, -1):
-        r.core.removeAssembliesInRing(ring, cs)
 
 
 class ReactorTests(unittest.TestCase):
@@ -867,7 +778,8 @@ class HexReactorTests(ReactorTests):
         for b in self.r.core.getBlocks():
             b.p.mgFlux = range(5)
             b.p.adjMgFlux = range(5)
-        with directoryChangers.TemporaryDirectoryChanger(root=THIS_DIR):
+
+        with directoryChangers.TemporaryDirectoryChanger(root=_THIS_DIR):
             self.r.core.saveAllFlux()
 
     def test_getFluxVector(self):
