+++ armi/utils/tests/test_utils.py
@@ -20,8 +20,8 @@ import unittest
 import numpy as np
 
 from armi import utils
-from armi.reactor.tests.test_reactors import loadTestReactor
 from armi.settings.caseSettings import Settings
+from armi.testing import loadTestReactor
 from armi.tests import mockRunLogs
 from armi.utils import (
     codeTiming,
